requirements.txt
flask==3.0.2
opencv-python==4.10.0.84
numpy==2.1.3
Pillow==11.0.0
werkzeug==3.0.1
imagehash==4.3.1
scikit-image==0.22.0
